[#3] refactor: MainPage 가독성 향상을 위한 컴포넌트 분리 작업
기존 MainTable에 있던 여러 기능들을 컴포넌트로 분리하는 추상화 작업 진행
MainPage에 들어 왔을때 테이블이 어떤 구성으로 되어있는지 알기 쉽게 하기 위함
